appointment-service/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
markdown==3.3.6
gunicorn==20.1.0
psycopg2-binary